refactor: :recycle: Modularize application into packages
- Separated logic from `main.go` into distinct packages to improve code organization and maintainability. - Added the following packages:   - `config`: handles application configurations.   - `controllers`: contains API business logic.   - `models`: defines data structures and models.   - `router`: manages server routing.   - `secrets`: handles secure credentials and sensitive variables. - Updated imports in `main.go` to reflect the new structure.
