+++ BackEnd/Domains/Auth/Login/config/cors.go
@@ -1,0 +1,16 @@
+package config
+
+import (
+	"github.com/gin-contrib/cors"
+	"github.com/gin-gonic/gin"
+)
+
+func GetCORSConfig() gin.HandlerFunc {
+	corsConfig := cors.New(cors.Config{
+		AllowOrigins:     []string{"http://localhost:3000"},
+		AllowMethods:     []string{"GET", "POST"},
+		AllowHeaders:     []string{"Origin", "Content-Type", "Authorization"},
+		AllowCredentials: true,
+	})
+	return corsConfig
+}
