+++ BackEnd/Domains/Auth/Login/go.mod
@@ -1,4 +1,4 @@
-module login-rest-api
+module login-service
 
 go 1.23.4
