+++ BackEnd/Domains/Auth/Login/models/claims.go
@@ -1,0 +1,9 @@
+package models
+
+import "github.com/golang-jwt/jwt/v5"
+
+type Claims struct {
+	jwt.RegisteredClaims
+	UserID string `json:"user_id"`
+	Role   string `json:"role"`
+}
