+++ BackEnd/Domains/Auth/Login/models/login.go
@@ -1,0 +1,6 @@
+package models
+
+type Login struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
